feat: adjust the application to operate on json
